refactor(test-framework): Remove unused abstractions
