+++ src/TestFramework/PhpUnit/Config/Builder/InitialConfigBuilder.php
@@ -35,7 +35,6 @@ declare(strict_types=1);
 
 namespace Infection\TestFramework\PhpUnit\Config\Builder;
 
-use Infection\TestFramework\Config\InitialConfigBuilder as ConfigBuilder;
 use Infection\TestFramework\PhpUnit\Adapter\PhpUnitAdapter;
 use Infection\TestFramework\PhpUnit\Config\XmlConfigurationManipulator;
 use Infection\TestFramework\PhpUnit\Config\XmlConfigurationVersionProvider;
@@ -48,7 +47,7 @@ use Webmozart\Assert\Assert;
 /**
  * @internal
  */
-final readonly class InitialConfigBuilder implements ConfigBuilder
+final readonly class InitialConfigBuilder
 {
     private string $originalXmlConfigContent;
