+++ tests/phpunit/Fixtures/TestFramework/FakeAwareAdapter.php
@@ -4,30 +4,12 @@ declare(strict_types=1);
 
 namespace Infection\Tests\Fixtures\TestFramework;
 
-use ErrorException;
 use Infection\AbstractTestFramework\MemoryUsageAware;
-use Infection\TestFramework\AbstractTestFrameworkAdapter;
-use Infection\Tests\UnsupportedMethod;
 
-class FakeAwareAdapter extends AbstractTestFrameworkAdapter implements MemoryUsageAware
+final class FakeAwareAdapter extends DummyTestFrameworkAdapter implements MemoryUsageAware
 {
     public function __construct(private readonly float $memoryLimit)
     {
-    }
-
-    public function hasJUnitReport(): bool
-    {
-        return false;
-    }
-
-    public function testsPass(string $output): bool
-    {
-        throw UnsupportedMethod::method(self::class, __FUNCTION__);
-    }
-
-    public function getName(): string
-    {
-        throw UnsupportedMethod::method(self::class, __FUNCTION__);
     }
 
     /**
